Conflicts:
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataServiceImpl.java
	service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/PeriodServiceImpl.java
	src/main/resources/form_template/ndfl/primary_rnu_ndfl/v2016/script.groovy
	src/main/resources/form_template/ndfl/report_2ndfl_1/v2016/script.groovy
	src/main/resources/form_template/ndfl/report_6ndfl/v2016/script.groovy

+++ service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DeclarationDataServiceImpl.java
@@ -13,6 +13,7 @@ import com.aplana.sbrf.taxaccounting.model.log.LogEntry;
 import com.aplana.sbrf.taxaccounting.model.log.LogLevel;
 import com.aplana.sbrf.taxaccounting.model.log.Logger;
 import com.aplana.sbrf.taxaccounting.model.refbook.RefBook;
+import com.aplana.sbrf.taxaccounting.model.util.Pair;
 import com.aplana.sbrf.taxaccounting.refbook.RefBookDataProvider;
 import com.aplana.sbrf.taxaccounting.refbook.RefBookFactory;
 import com.aplana.sbrf.taxaccounting.service.*;
@@ -82,6 +83,8 @@ public class DeclarationDataServiceImpl implements DeclarationDataService {
             "консолидация).";
     private static final String CALCULATION_NOT_TOPICAL_SUFFIX = " Для коррекции консолидированных данных необходимо нажать на кнопку \"Рассчитать\"";
     private static final int DEFAULT_TF_FILE_TYPE_CODE = 1;
+
+    final static List<DeclarationDataReportType> reportTypes = Collections.unmodifiableList(Arrays.asList(DeclarationDataReportType.ACCEPT_DEC, DeclarationDataReportType.CHECK_DEC, DeclarationDataReportType.XML_DEC, DeclarationDataReportType.IMPORT_TF_DEC, DeclarationDataReportType.DELETE_DEC));
 
     @Autowired
     private DeclarationDataDao declarationDataDao;
@@ -231,7 +234,7 @@ public class DeclarationDataServiceImpl implements DeclarationDataService {
                 declarationDataAccessService.checkEvents(userInfo, declarationTemplateId, departmentReportPeriod,
                         FormDataEvent.CREATE, logger);
                 if (logger.containsLevel(LogLevel.ERROR)) {
-                    throw new ServiceLoggerException(("Уведомление не создано"), logEntryService.save(logger.getEntries()));
+                    throw new ServiceLoggerException(("Налоговая форма не создана"), logEntryService.save(logger.getEntries()));
                 }
 
                 DeclarationData newDeclaration = new DeclarationData();
@@ -367,31 +370,55 @@ public class DeclarationDataServiceImpl implements DeclarationDataService {
     @Override
     @Transactional(readOnly = false)
     public void delete(long id, TAUserInfo userInfo) {
+        delete(id, userInfo, true);
+    }
+
+    @Override
+    @Transactional(readOnly = false)
+    public void delete(long id, TAUserInfo userInfo, boolean createLock) {
         LockData lockData = lockDataService.getLock(generateAsyncTaskKey(id, DeclarationDataReportType.XML_DEC));
         LockData lockDataAccept = lockDataService.getLock(generateAsyncTaskKey(id, DeclarationDataReportType.ACCEPT_DEC));
         LockData lockDataCheck = lockDataService.getLock(generateAsyncTaskKey(id, DeclarationDataReportType.CHECK_DEC));
-        if (lockData == null && lockDataAccept == null && lockDataCheck == null) {
-            declarationDataAccessService.checkEvents(userInfo, id, FormDataEvent.DELETE);
-            DeclarationData declarationData = declarationDataDao.get(id);
+        LockData lockDataDelete = null;
+        if (lockData == null && lockDataAccept == null && lockDataCheck == null && createLock) {
+            lockDataDelete = lockDataService.lock(generateAsyncTaskKey(id, DeclarationDataReportType.DELETE_DEC), userInfo.getUser().getId(),
+                    getDeclarationFullName(id, DeclarationDataReportType.DELETE_DEC));
+        }
+        if (lockData == null && lockDataAccept == null && lockDataCheck == null && lockDataDelete == null) {
+            try {
+                declarationDataAccessService.checkEvents(userInfo, id, FormDataEvent.DELETE);
+                DeclarationData declarationData = declarationDataDao.get(id);
 
-            Logger logger = new Logger();
-            declarationDataScriptingService.executeScript(userInfo,
-                    declarationData, FormDataEvent.DELETE, new Logger(), null);
+                Logger logger = new Logger();
+                declarationDataScriptingService.executeScript(userInfo,
+                        declarationData, FormDataEvent.DELETE, new Logger(), null);
 
-            // Проверяем ошибки
-            if (logger.containsLevel(LogLevel.ERROR)) {
-                throw new ServiceLoggerException(
-                        "Найдены ошибки при выполнении удаления налоговой формы",
-                        logEntryService.save(logger.getEntries()));
+                // Проверяем ошибки
+                if (logger.containsLevel(LogLevel.ERROR)) {
+                    throw new ServiceLoggerException(
+                            "Найдены ошибки при выполнении удаления налоговой формы",
+                            logEntryService.save(logger.getEntries()));
+                }
+
+                deleteReport(id, userInfo, false, TaskInterruptCause.DECLARATION_DELETE);
+                declarationDataDao.delete(id);
+
+                auditService.add(FormDataEvent.DELETE , userInfo, declarationData, "Налоговая форма удалена", null);
+            } finally {
+                if (createLock) {
+                    lockDataService.unlock(generateAsyncTaskKey(id, DeclarationDataReportType.DELETE_DEC), userInfo.getUser().getId());
+                }
             }
-
-            deleteReport(id, userInfo, false, TaskInterruptCause.DECLARATION_DELETE);
-            declarationDataDao.delete(id);
-
-            auditService.add(FormDataEvent.DELETE , userInfo, declarationData, "Налоговая форма удалена", null);
         } else {
-            if (lockData == null) lockData = lockDataAccept;
-            if (lockData == null) lockData = lockDataCheck;
+            if (lockData == null) {
+                lockData = lockDataAccept;
+            }
+            if (lockData == null) {
+                lockData = lockDataCheck;
+            }
+            if (lockData == null) {
+                lockData = lockDataDelete;
+            }
             Logger logger = new Logger();
             TAUser blocker = taUserService.getUser(lockData.getUserId());
             logger.error("Текущая налоговая форма не может быть удалена, т.к. пользователем \"%s\" в \"%s\" запущена операция \"%s\"", blocker.getName(), SDF_DD_MM_YYYY_HH_MM_SS.get().format(lockData.getDateLock()), lockData.getDescription());
@@ -1366,7 +1393,7 @@ public class DeclarationDataServiceImpl implements DeclarationDataService {
                                 : "");
             case DELETE_DEC:
                 return String.format(LockData.DescriptionTemplate.DECLARATION_TASK.getText(),
-                        "налоговой формы",
+                        getTaskName(ddReportType, TaxType.NDFL),
                         reportPeriod.getReportPeriod().getName() + " " + reportPeriod.getReportPeriod().getTaxPeriod().getYear(),
                         reportPeriod.getCorrectionDate() != null
                                 ? " с датой сдачи корректировки " + sdf.get().format(reportPeriod.getCorrectionDate())
@@ -1511,6 +1538,7 @@ public class DeclarationDataServiceImpl implements DeclarationDataService {
             case EXCEL_DEC:
             case TO_CREATE_DEC:
             case XML_DEC:
+            case DELETE_DEC:
                 return String.format(ddReportType.getReportType().getDescription(), taxType.getDeclarationShortName());
             case PDF_DEC:
                 return String.format(ddReportType.getReportType().getDescription(), "");
@@ -1837,4 +1865,17 @@ public class DeclarationDataServiceImpl implements DeclarationDataService {
         }
         return toReturn;
     }
+
+    @Override
+    public Map<DeclarationDataReportType, LockData> getLockTaskType(long declarationDataId) {
+        Map<DeclarationDataReportType, LockData> result = new HashMap<DeclarationDataReportType, LockData>();
+        for (DeclarationDataReportType reportType : reportTypes) {
+            LockData lockData = lockDataService.getLock(generateAsyncTaskKey(declarationDataId, reportType));
+            if (lockData != null) {
+                result.put(reportType, lockData);
+            }
+        }
+        return result;
+    }
+
 }
